Squashed commit of the following:
commit ed3b8c02d8938b24566bd4be4ce7589f80019930
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Mon Nov 17 20:02:28 2025 -0800

    Added `-based` to the tested flags

commit c33747df22a17da34ce88c2451cb29397887e629
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Mon Nov 17 19:59:17 2025 -0800

    Build script stuff + Linker bug fixes

commit becd1d2138e639b52ffd57df1eab66bb67850918
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Nov 16 19:10:55 2025 -0800

    Mini-fixes

commit dd66b4567e96d103b9665e5b90b64fb0d474e26b
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Nov 16 19:05:11 2025 -0800

    Got tests passing again

commit dfb7a8224082d3618fa4d96ee0f2f8f31eb8c7a8
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Nov 16 16:35:00 2025 -0800

    Lexer gen rewrite mostly there

commit d389b70d290d3190ba1b019c7cc1afec1fdd9e2b
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Fri Nov 7 15:58:42 2025 -0800

    Revert-ish

commit 6576469440ec61187e05d3edb9e657f1f66fd50f
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Tue Nov 4 21:56:00 2025 -0800

    effortys

commit e33342629c90e55661540814f7e386a8db07ac52
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Tue Oct 21 00:15:33 2025 -0700

    Update lexgen.lua

commit ad9f906719bf0997fbf5a07d26527c6e1f96a69a
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Mon Oct 20 18:27:59 2025 -0700

    Update lexgen.lua

commit 14f8d515ac7a21ee66d8074e77f296e2dd55e299
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Oct 19 17:21:36 2025 -0700

    CI update

commit ef123dcd0fb4c1a1ba5a9d22fb96918027d2b0cd
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Oct 19 17:18:33 2025 -0700

    Getting things compiling again

commit d0b81e529756751057ccb119dd5fa80e3e0091ec
Merge: 64d36c5 009b265
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sat Oct 18 14:35:54 2025 -0700

    Merge branch 'master' into dev

commit 64d36c582f9b8ecc4b57637ce9de33d7874d7bbb
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Mon Oct 13 19:14:51 2025 -0700

    GH

commit c227906e948bae1e8fe4a8166520366aaa036786
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sun Oct 12 13:28:05 2025 -0700

    Progres

commit 4237d7e3a555b9c1e8f521bd620f876fe29e0f40
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Sat Oct 11 16:53:00 2025 -0700

    More Go features

commit 5d94569f6e295e58bcab927cd2663fabc799ac2c
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Fri Oct 10 18:35:55 2025 -0700

    Doing it

commit 42eeb28d3cb90d627d163931f4b3c6fbfbae591b
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Fri Oct 10 16:29:02 2025 -0700

    It's lowkey doing what it do

commit 6d8662db16e07f6eb4a60406c40d3edfaf39f660
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Thu Oct 9 00:34:19 2025 -0700

    Work on the GC remap, tweak of RA costs

commit 9e75451b5f35f0fbcc4c1de48aa74d0966fae9ed
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Wed Oct 8 22:03:50 2025 -0700

    Almost!!!

commit 8c3436d16039210f291771ce5fd87437c275ffd9
Author: Yasser Arguelles Snape <yasserarguelles2506@gmail.com>
Date:   Wed Oct 8 17:42:59 2025 -0700

    [go] Improving GC support

+++ include/tb_formats.h
@@ -35,7 +35,7 @@ typedef struct {
     TB_ObjectRelocType type;
     uint32_t symbol_index;
     size_t virtual_address;
-    size_t addend;
+    int32_t addend;
 } TB_ObjectReloc;
 
 typedef enum {
